% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{combinations}
\alias{combinations}
\title{List of triads to be ranked for parentage assignment.}
\format{
An object of class \code{data.frame} with 24192 rows and 24 columns.
}
\usage{
combinations
}
\description{
An example list of how the output from the hothiphop function looks like.
Listed are all possible triad combinations for 8 offspring from the 2018 cohort
This dataframe is also used to generate the vignette.
See the help file of the hothiphop function for an explanation of columns.
}
\keyword{datasets}
