% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{genotypes}
\alias{genotypes}
\title{Genotypes of individuals scored at different loci}
\format{
A data frame with 1407 rows (individuals) and 1376 variables (loci)
}
\source{
Cockburn et al. (2020) HIPHOP: improved paternity assignment among
close relatives using a simple exclusion method for biallelic markers.
Molecular Ecology Resources, in revision.
}
\usage{
genotypes
}
\description{
A dataset containing genotypes of 1407 superb fairy wrens (Malurus cyaneus)
from five cohorts (breeding seasons 2014-2018) from the Australian National Botanic Garden.
Each individual (rows) is scored at 1376 loci (columns) with the scores meaning:
0: homozygotes at common allele; 1: homozygote at rare allele and
2: heterozygotes; NA: locus could not be scored.
}
\author{
Andrew Cockburn, \email{andrew.cockburn@anu.edu.au}
}
\keyword{datasets}
