% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qc.R
\name{p12}
\alias{p12}
\title{Building a Model with Top Ten Features}
\usage{
p12(matI, x, sig = 1, col = col, ...)
}
\arguments{
\item{sig}{an optional sig cutoff}

\item{mat}{a matrix of values}

\item{sig}{an optional color}
}
\value{
plot
}
\description{
This function plots the log ratios of a matrix
}
\details{
This function runs a univariate regression of y on each predictor in x and
calculates the p-value indicating the significance of the association. The
final set of 10 predictors is the taken from the 10 smallest p-values.
}
\author{
giaever
}

